requirements.txt
Flask==3.1.1
flask-cors==6.0.1
mysql-connector-python==9.3.0
pymongo==4.6.1
PyJWT==2.8.0
cryptography==42.0.5
Werkzeug==3.1.3
gunicorn==21.2.0
